[Render] Change code to handle shaders
